Make new round template sets
