+++ src/app/round-templates/round-template.spec.ts
@@ -1,0 +1,39 @@
+import { defaultSets } from './round-template';
+
+describe('Sets of Round Templates', () => {
+  defaultSets.forEach(({name: setName, templates}) => {
+    describe(setName, () => {
+      it('Contains at least one round', () => {
+        expect(templates.length).toBeGreaterThan(0);
+      });
+
+      templates.forEach(({name, flowerSpecies, startTime, endTime, tickSpeed, bees}) => {
+        describe(name, () => {
+          it('has either 8 or 16 flowers', () => {
+            expect([8, 16]).toContain(flowerSpecies.length);
+          });
+
+          if (bees) {
+            describe('if there are bees', () => {
+              it('has at least 1 bee', () => {
+                expect(bees.length).toBeGreaterThanOrEqual(1);
+              });
+              it('bee weights add to 1', () => {
+                expect(bees.reduce((prev, curr) => prev + curr.weight, 0)).toBeCloseTo(1, 10);
+              });
+            });
+          }
+
+          it('has a reasonable tickSpeed', () => {
+            expect(tickSpeed).toBeGreaterThanOrEqual(100);
+            expect(tickSpeed).toBeLessThan(100000);
+          });
+
+          it('has chronological startTime and endTime', () => {
+            expect(startTime.time).toBeLessThan(endTime.time);
+          });
+        });
+      });
+    });
+  });
+});
