+++ tsconfig.json
@@ -15,7 +15,7 @@
       "node_modules/@types"
     ],
     "lib": [
-      "es2018",
+      "es2019",
       "dom"
     ],
     "resolveJsonModule": true
